Refactor Belgium to use Country class
Also modifies the two population partitions to generate full derived
metrics in the way NI is done.
